dedup test classes and auto complete full address
