+++ tests/Db/CollectedAddressMapperTest.php
@@ -22,53 +22,84 @@
 
 namespace OCA\Mail\Tests\Db;
 
-use OC;
 use OC\AppFramework\Db\Db;
-use OCA\Mail\Db\CollectedAddress;
+use Test\TestCase;
 use OCA\Mail\Db\CollectedAddressMapper;
-use OCP\IDBConnection;
-use PHPUnit_Framework_TestCase;
+use OCA\Mail\Db\CollectedAddress;
 
-class CollectedAddressMapperTest extends PHPUnit_Framework_TestCase {
+/**
+ * Class CollectedAddressMapperTest
+ *
+ * @group DB
+ *
+ * @package OCA\Mail\Tests\Db
+ */
+class CollectedAddressMapperTest extends TestCase {
 
-	/** @var IDBConnection */
+	/** @var \OCP\IDBConnection */
 	private $db;
-	/** @var string */
 	private $userId = 'testuser';
+
 	/** @var CollectedAddressMapper */
 	private $mapper;
+
 	/** @var CollectedAddress */
 	private $address1;
+
 	/** @var CollectedAddress */
 	private $address2;
+
+	/** @var CollectedAddress */
+	private $address3;
 
 	protected function setUp() {
 		parent::setUp();
 
-		$this->db = OC::$server->getDatabaseConnection();
+		$this->db = \OC::$server->getDatabaseConnection();
 		$this->mapper = new CollectedAddressMapper(new Db($this->db));
 
 		$this->address1 = new CollectedAddress();
 		$this->address1->setEmail('user1@example.com');
+		$this->address1->setDisplayName('User 1');
 		$this->address1->setUserId($this->userId);
 
 		$this->address2 = new CollectedAddress();
 		$this->address2->setEmail('user2@example.com');
+		$this->address2->setDisplayName('User 2');
 		$this->address2->setUserId($this->userId);
 
-		$sql = 'INSERT INTO *PREFIX*mail_collected_addresses (`email`, `user_id`) VALUES (?, ?)';
+		$this->address3 = new CollectedAddress();
+		$this->address3->setEmail('"User 3" <user3@domain.com>');
+		$this->address3->setDisplayName('User 3');
+		$this->address3->setUserId($this->userId);
+
+		$qb = $this->db->getQueryBuilder();
+		$sql = 'INSERT INTO *PREFIX*mail_collected_addresses (`email`, `display_name`, `user_id`) VALUES (?, ?, ?)';
 		$stmt = $this->db->prepare($sql);
+
+		// Empty DB
+		$qb = $this->db->getQueryBuilder();
+		$qb->delete($this->mapper->getTableName());
+		$qb->execute();
 
 		$stmt->execute([
 			$this->address1->getEmail(),
+			$this->address1->getDisplayName(),
 			$this->address1->getUserId(),
 		]);
 		$this->address1->setId($this->db->lastInsertId('PREFIX*mail_collected_addresses'));
 		$stmt->execute([
 			$this->address2->getEmail(),
+			$this->address2->getDisplayName(),
 			$this->address2->getUserId(),
 		]);
 		$this->address2->setId($this->db->lastInsertId('PREFIX*mail_collected_addresses'));
+		$stmt->execute([
+			$this->address3->getEmail(),
+			$this->address3->getDisplayName(),
+			$this->address3->getUserId(),
+		]);
+		$this->address3->setId($this->db->lastInsertId('PREFIX*mail_collected_addresses'));
 	}
 
 	protected function tearDown() {
@@ -82,12 +113,15 @@ class CollectedAddressMapperTest extends PHPUnit_Framework_TestCase {
 		if (!empty($this->address2)) {
 			$stmt->execute([$this->address2->getId()]);
 		}
+		if (!empty($this->address3)) {
+			$stmt->execute([$this->address3->getId()]);
+		}
 	}
 
 	public function matchingData() {
 		return [
-			['user1@example.com', ['user1@example.com']],
-			['ser', ['user1@example.com', 'user2@example.com']],
+				['user1@example.com', ['user1@example.com']],
+				['examp', ['user1@example.com', 'user2@example.com']],
 		];
 	}
 
@@ -109,8 +143,8 @@ class CollectedAddressMapperTest extends PHPUnit_Framework_TestCase {
 
 	public function existsData() {
 		return [
-			['user1@example.com', true],
-			['user3@example.com', false],
+				['user1@example.com', true],
+				['user3@example.com', false],
 		];
 	}
 
@@ -121,6 +155,24 @@ class CollectedAddressMapperTest extends PHPUnit_Framework_TestCase {
 		$actual = $this->mapper->exists($this->userId, $email);
 
 		$this->assertSame($expected, $actual);
+	}
+
+	public function testGetTotal() {
+		$total = $this->mapper->getTotal();
+
+		$this->assertSame(3, $total);
+	}
+
+	public function testGetChunk() {
+		$chunk = $this->mapper->getChunk();
+
+		$this->assertCount(3, $chunk);
+	}
+
+	public function testGetChunkWithOffset() {
+		$chunk = $this->mapper->getChunk($this->address2->getId());
+
+		$this->assertCount(2, $chunk);
 	}
 
 }
